Seperates visualization code from the core drive.c
Changes
- Remove GIF generation code from drive.c
- Improved load_weights to auto-detect file size
- view flag
- random maps if map-name not passed
- policy-name flag to make videos for a particular policy. Saved in the policy directoryq
