+++ pufferlib/extensions/puffernet.h
@@ -50,7 +50,19 @@ void _load_weights(const char* filename, float* weights, size_t num_weights) {
     }
 }
 
-Weights* load_weights(const char* filename, size_t num_weights) {
+Weights* load_weights(const char* filename) {
+    FILE* file = fopen(filename, "rb");
+    if (!file) {
+        perror("Error opening weights file");
+        return NULL;
+    }
+    fseek(file, 0, SEEK_END);
+    size_t file_size = ftell(file);
+    fclose(file);
+
+    size_t num_weights = file_size / sizeof(float);
+    printf("Loading %zu weights from %s\n", num_weights, filename);
+
     Weights* weights = calloc(1, sizeof(Weights) + num_weights*sizeof(float));
     weights->data = (float*)(weights + 1);
     _load_weights(filename, weights->data, num_weights);
